requirements.txt
Package       Version
------------- -------
Pillow        7.0.0  
pip           20.0.2 
pkg-resources 0.0.0  
setuptools    39.0.1 
